--- frontend/src/api/api.ts
@@ -1,10 +1,0 @@
-import axios from 'axios';
-
-
-export default axios.create({
-  baseURL: process.env.REACT_APP_API_URL,
-  timeout: 1000,
-  headers: {
-    'x-auth-token': localStorage.getItem('token')
-  },
-});
