+++ frontend/src/components/ManageTeams/ManageTeams.stories.tsx
@@ -3,9 +3,10 @@ import { Story, Meta } from '@storybook/react/types-6-0';
 import ManageTeams , { ManageTeamsProps } from './ManageTeams';
 import { store } from '../../app/store';
 import { Provider } from 'react-redux';
-import getTeams from '../../api/getTeams';
+import getTeams from '../../api/Teams.service';
 import darkTheme from '../../theme/customMaterialTheme';
 import { ThemeProvider } from '@material-ui/styles';
+import TeamService from '../../api/Teams.service';
 
 
 export default {
@@ -40,9 +41,10 @@ SampleFakeManageTeams.args = {
 
 const token = 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI2MDRjNjQyZTUzNDdhZDE5ZDRmOWE0MzciLCJ0eXBlIjozLCJpYXQiOjE2MTY4NzIzNTcsImV4cCI6MTYxNjk1ODc1N30.RX1EzN6tHmdMFhDWtm6TmQPFzML6min3e_11RH3B6GA';
 localStorage.setItem('token', token);
+const api = new TeamService();
 
 export const SampleManageTeams = Template.bind({});
 SampleManageTeams.args = {
-  getTeams: getTeams,
+  getTeams: api.getTeams,
   onClickAdd: () => console.log('Clicked Add'),
 };
