+++ frontend/src/components/ReusableTable/ReusableTable.stories.tsx
@@ -3,8 +3,7 @@ import { Story, Meta } from '@storybook/react/types-6-0';
 import ReusableTable , { ReusableTableProps } from './ReusableTable';
 import { store } from '../../app/store';
 import { Provider } from 'react-redux';
-
-import getTodos from '../../api/getTodos';
+import TeamService from '../../api/Teams.service';
 
 export default {
   title: 'ReusableTable component',
@@ -17,10 +16,12 @@ const Template: Story<ReusableTableProps> = (args) => (
   </Provider>
 );
 
+const api = new TeamService();
+
 export const SampleReusableTable = Template.bind({});
 SampleReusableTable.args = {
   name: 'My Table',
-  columns: [{field: 'id', width: 100}, {field: 'title', width: 400}],
-  getData: getTodos,
+  columns: [{field: 'id', width: 200}, {field: 'name', width: 200}, {field: 'surname', width: 200}, {field: 'courseName', width: 200}],
+  getData: api.getTeams,
   onRowClick: (params, e) => console.log(params.row),
 };
