+++ frontend/src/models/Course.model.ts
@@ -1,0 +1,4 @@
+export interface CourseData {
+    _id: string;
+    name: string;
+}
